backend/requirements.txt
asgiref==3.8.1
boto3==1.36.26
botocore==1.36.26
Django==5.1.5
jmespath==1.0.1
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
